requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
jalali_core==1.0.0
jdatetime==5.0.0
mysql-connector-python==8.4.0
pyTelegramBotAPI==4.18.1
requests==2.32.0
urllib3==2.2.1

